Doors are now interactable.
All 30 types of doors are interactable, meaning open and close. Added a script that determines state and projects text on the screen to prompt if the player is within the collider. If the text is being displayed and the player hits the `R` key, then the door interacts (opens if closed and closes if opens).

The textfield has to be manually added per door and can't be saved as prefab (it erases afterwards). If a door is NOT supposed to open, AKA a door with bars on the door - it will not have a text field.

If the door IS supposed to open it will have a text field.